Convert all the library code in Kotlin
The project migrated to androidx, andoid's newest support library, for better future support of all android's classes.

All code of the module geofirestore is converted to Kotlin except the file GeoQuery.java, this help reduce the boilerplate code typical of java and allows the use of kotlin's new features.
The structure of the library remains unchanged, methods and variables do not change their name; the get and set methods for the variables are replaced by the variable itself in accordance with the rules for accessing kotlin attributes.

Changed the geo hash generation algorithm due to a discrepancy of the result produced by the library and by some famous online geohashing tools (such as: http://geohash.co); now the hash produced for the same pair of coordinates is the same.
The new algorithm works like this; there are two nested loop:
   - the inner one cycle every bit from 0 to 4 considering if it's in an even position
     if so we calculate the value based on the longitude by finding the error (mean of min longitude end max longitude), then the value is added and the range of longitude is updated with the recently calculated error
     otherwise we calculate the value based on the latitude doing the same calculations
     at the end the value is converted to the corresponding Base32 char using the Base32Utils method
   - the outer loop repeat the value calculation until we obtain a word of length precision

GeoFirestore.kt includes method setLocation, getLocation, removeLocation that allow the use of lambdas instead of implementing CompletionListener and LocationListener interfaces, so the kotlin's users can use the library with less boilerplate code.