Continue migration to Kotlin
Refactored the package structure of the library in order to simplify the codebase and group filed with similar scope (all the listeners in callbacks package, extension functions in extensions package).
Updated some of the android and kotiln external libraries to the newer version.

The MainActivity.java file is converted to kotlin to take advantage of its features and step up the full conversion of the codebase.
Base32Utils, Constants, GeoUtils are now kotlin's object because they contains all static methods previously present in the companion object.
The code responsible of generating the hash in GeoHash.kt is moved in a separate method.

SingleGeoQuery class is like GeoQuery but instead of listening the changes of all the documents in the radius it return "one-shot" all the documents in the radius at the time it's called.
Note: this class was created for personal use in al project but it's keep becouse it can bee handy for someone else.

The GeoFirestoreExt file redefine all the GeoFirebase's public methods (getLocation, removeLocation, setLocation) but with a lambda function instead of the callback as parameters; this is done for group all the kotlin's extension, but we can reintroduce those methods in the main class is we want.