Continue migration to Kotlin
Updated some external libraries in app's and geofirestore's build.gradle files; kotlin is updated from 1.3.21 to 1.3.31.
In the app module the class MainActivity.java is converted to MainActivity.kt.

Created a package callbacks and moved in it EventListenerBridge.kt, GeoQueryEventListener.kt, GeoQueryDataEventListener.kt, SingleGeoQueryDataEventListener.kt

Base32Utils.kt, Constants.kt, GeoUtils.kt are now kotlin's object because they contains all static methods.

The code responsible of generating the hash in GeoHash.kt is moved in a separate method.
GeoHashQuery's Util subclass is now a kotlin's object.

Added SingleGeoQuery.kt, this class is like a GeoQuery but instead of listening all the documents in the radius it works "one-shot" returning all the documents in the radius at the time it's called. Note: this class was created for personal use.
SingleGeoQueryDataEventListener.kt contain the callback interface for SingleGeoQuery

the package extension is created to hold files with extension functions like CollectionExtension.kt and GeoFirestoreExt.kt, the second file declare all the public method of GeoFirestore but with kotlin's lambda functions instead of the interfaces.