Refactor some interfaces names
The name Listener is used when we want to receive continuous updates, while the Callback name is used when we want to receive a reply only once; based on this nomenclature the CompletionListener and the SingleGeoQueryDataEventListener interfaces have been renamed to CompletionCallback and SingleGeoQueryDataEventCallback.

Moved the check for empty strings from the GeoHash constructor to the method isValidBase32String in Base32Utils because is more logical for him to check the string and return if it's empty.