Modify the README to add a CHANGELOG
Specified how to import the library to the right version with Gradle or Maven.
All code examples are converted to Kotlin.
The API Reference & Documentation section removed to resolve issue # 25, the link is corrupt.
Added a section showing all the apps that use GeoFirestore and a section that refers to CHANGELOG